+++ src/app/components/lb/lb.component.ts
@@ -1,14 +1,14 @@
 import { Component, Input } from '@angular/core';
 
-import { LbData } from '../../models/parsed-elements';
+import { Lb } from '../../models/evt-models';
 import { register } from '../../services/component-register.service';
 
-@register(LbData)
+@register(Lb)
 @Component({
   selector: 'evt-lb',
   templateUrl: './lb.component.html',
   styleUrls: ['./lb.component.scss'],
 })
 export class LbComponent {
-  @Input() data: LbData;
+  @Input() data: Lb;
 }