+++ src/app/services/entities-select.service.ts
@@ -3,7 +3,7 @@ import { Subject } from 'rxjs';
 import { shareReplay } from 'rxjs/operators';
 import { AppConfig } from '../app.config';
 import { EntitiesSelectItem } from '../components/entities-select/entities-select.component';
-import { AttributesData } from '../models/evt-models';
+import { Attributes } from '../models/evt-models';
 
 @Injectable({
   providedIn: 'root',
@@ -23,7 +23,7 @@ export class EntitiesSelectService {
       .map(i => i.replace(/(\[|\]|\')/g, '').split('=')).map(i => ({ key: i[0], value: i[1] }));
   }
 
-  public matchClassAndAttributes(valueForCheck: string, attributesToCheck: AttributesData, classToCheck: string) {
+  public matchClassAndAttributes(valueForCheck: string, attributesToCheck: Attributes, classToCheck: string) {
     return valueForCheck.split(',')
       .some(v => this.matchClass(v, classToCheck) && this.matchAttributes(v, attributesToCheck));
   }
@@ -32,11 +32,11 @@ export class EntitiesSelectService {
     return classToCheck === this.getClassNameFromValue(classForCheck);
   }
 
-  public matchAttributes(attributesForCheck: string, attributesToCheck: AttributesData) {
+  public matchAttributes(attributesForCheck: string, attributesToCheck: Attributes) {
     return this.getAttributesFromValue(attributesForCheck).every(a => attributesToCheck[a.key] === a.value);
   }
 
-  public getHighlightColor(attributesToCheck: AttributesData, classNameToCheck: string, selectedItems?: EntitiesSelectItem[]) {
+  public getHighlightColor(attributesToCheck: Attributes, classNameToCheck: string, selectedItems?: EntitiesSelectItem[]) {
     const entitiesSelectItems = AppConfig.evtSettings.edition.entitiesSelectItems
       .reduce((i: EntitiesSelectItem[], g) => i.concat(g.items), [])
       .reduce((x: EntitiesSelectItem[], y) => {