+++ src/app/services/xml-parsers/named-entities-parser.service.ts
@@ -1,14 +1,15 @@
 import { Injectable } from '@angular/core';
-import { AttributesMap } from 'ng-dynamic-component';
+import { parse } from '.';
 import { AppConfig } from '../../app.config';
 import {
-  Description, NamedEntitiesList, NamedEntity, NamedEntityInfo,
-  NamedEntityLabel, NamedEntityOccurrence, NamedEntityOccurrenceRef, NamedEntityType, PageData, Relation, XMLElement,
+  NamedEntitiesList, NamedEntity, NamedEntityOccurrence, NamedEntityOccurrenceRef, NamedEntityType, Page, XMLElement,
 } from '../../models/evt-models';
 import { isNestedInElem, xpath } from '../../utils/dom-utils';
 import { Map } from '../../utils/js-utils';
 import { replaceNewLines } from '../../utils/xml-utils';
-import { GenericParserService } from './generic-parser.service';
+import { AttributeMapParser, ElementParser } from './basic-parsers';
+import { RelationParser } from './named-entity-parsers';
+import { createParser } from './parser-models';
 
 @Injectable({
   providedIn: 'root',
@@ -22,11 +23,6 @@ export class NamedEntitiesParserService {
     events: 'listEvent',
     occurrences: 'persName[ref], placeName[ref], orgName[ref], geogName[ref], event[ref]',
   };
-
-  constructor(
-    private genericParserService: GenericParserService,
-  ) {
-  }
 
   public parseLists(document: XMLElement) {
     const listsToParse = this.getListsToParseTagNames();
@@ -43,6 +39,7 @@ export class NamedEntitiesParserService {
   }
 
   private parseList(list: XMLElement) {
+    const attributeParser = createParser(AttributeMapParser, parse);
     const parsedList: NamedEntitiesList = {
       type: NamedEntitiesList,
       id: list.getAttribute('xml:id') || xpath(list),
@@ -53,8 +50,10 @@ export class NamedEntitiesParserService {
       originalEncoding: list,
       relations: [],
       description: [],
-      attributes: this.parseAttributes(list),
+      attributes: attributeParser.parse(list),
     };
+
+    const relationParse = createParser(RelationParser, parse);
     list.childNodes.forEach((child: XMLElement) => {
       if (child.nodeType === 1) {
         switch (child.tagName.toLowerCase()) {
@@ -62,16 +61,16 @@ export class NamedEntitiesParserService {
             parsedList.label = replaceNewLines(child.textContent);
             break;
           case 'desc':
-            parsedList.description.push(this.genericParserService.parse(child));
+            parsedList.description.push(parse(child));
             break;
           case 'relation':
             if (this.neListsConfig.relations.enabled) {
-              parsedList.relations.push(this.parseRelation(child));
+              parsedList.relations.push(relationParse.parse(child));
             }
             break;
           case 'listrelation':
             if (this.neListsConfig.relations.enabled) {
-              child.querySelectorAll<XMLElement>('relation').forEach(r => parsedList.relations.push(this.parseRelation(r)));
+              child.querySelectorAll<XMLElement>('relation').forEach(r => parsedList.relations.push(relationParse.parse(r)));
             }
             break;
           default:
@@ -81,7 +80,7 @@ export class NamedEntitiesParserService {
               parsedList.content = parsedList.content.concat(parsedSubList.content);
               parsedList.relations = parsedList.relations.concat(parsedSubList.relations);
             } else {
-              parsedList.content.push(this.parseNamedEntity(child));
+              parsedList.content.push(parse(child) as NamedEntity);
             }
         }
       }
@@ -89,163 +88,6 @@ export class NamedEntitiesParserService {
     parsedList.label = parsedList.label || list.getAttribute('type') || `List of ${parsedList.namedEntityType}`;
 
     return parsedList;
-  }
-
-  private parseNamedEntity(xml: XMLElement): NamedEntity {
-    switch (xml.tagName) {
-      case 'person':
-        return this.parsePerson(xml);
-      case 'personGrp':
-        return this.parsePersonGroup(xml);
-      case 'place':
-        return this.parsePlace(xml);
-      case 'org':
-        return this.parseOrganization(xml);
-      case 'event':
-        return this.parseEvent(xml);
-      default:
-        console.error('Warning! Parser not implemented for this element in list', xml);
-
-        return;
-    }
-  }
-
-  private parseGenericEntity(xml: XMLElement): NamedEntity {
-    const elId = xml.getAttribute('xml:id') || xpath(xml);
-    const label = replaceNewLines(xml.textContent) || 'No info';
-    const entity: NamedEntity = {
-      type: NamedEntity,
-      id: elId,
-      sortKey: xml.getAttribute('sortKey') || (label ? label[0] : '') || xml.getAttribute('xml:id') || xpath(xml),
-      originalEncoding: xml,
-      label,
-      namedEntityType: this.getEntityType(xml.tagName),
-      content: Array.from(xml.children).map((subchild: XMLElement) => this.parseEntityInfo(subchild)),
-      attributes: this.parseAttributes(xml),
-    };
-
-    return entity;
-  }
-
-  private parsePerson(xml: XMLElement): NamedEntity {
-    const nameElement = xml.querySelector<XMLElement>('name');
-    const forenameElement = xml.querySelector<XMLElement>('forename');
-    const surnameElement = xml.querySelector<XMLElement>('surname');
-    const persNameElement = xml.querySelector<XMLElement>('persName');
-    const occupationElement = xml.querySelector<XMLElement>('occupation');
-    let label: NamedEntityLabel = 'No info';
-    if (persNameElement) {
-      label = replaceNewLines(persNameElement.textContent) || 'No info';
-    } else if (forenameElement || surnameElement) {
-      label += forenameElement ? `${replaceNewLines(forenameElement.textContent)} ` : '';
-      label += surnameElement ? `${replaceNewLines(surnameElement.textContent)} ` : '';
-    } else if (nameElement) {
-      label = replaceNewLines(nameElement.textContent) || 'No info';
-    } else {
-      label = replaceNewLines(xml.textContent) || 'No info';
-    }
-    label += occupationElement ? ` (${replaceNewLines(occupationElement.textContent)})` : '';
-
-    return {
-      ...this.parseGenericEntity(xml),
-      label,
-    };
-  }
-
-  private parsePersonGroup(xml: XMLElement): NamedEntity {
-    const role = xml.getAttribute('role');
-    let label: NamedEntityLabel = 'No info';
-    if (role) {
-      label = role.trim();
-    } else {
-      label = replaceNewLines(xml.textContent) || 'No info';
-    }
-
-    return {
-      ...this.parseGenericEntity(xml),
-      label,
-    };
-  }
-
-  private parsePlace(xml: XMLElement): NamedEntity {
-    const placeNameElement = xml.querySelector<XMLElement>('placeName');
-    const settlementElement = xml.querySelector<XMLElement>('settlement');
-    let label: NamedEntityLabel = 'No info';
-    if (placeNameElement) {
-      label = replaceNewLines(placeNameElement.textContent) || 'No info';
-    } else if (settlementElement) {
-      label = replaceNewLines(settlementElement.textContent) || 'No info';
-    }
-
-    return {
-      ...this.parseGenericEntity(xml),
-      label,
-    };
-  }
-
-  private parseOrganization(xml: XMLElement): NamedEntity {
-    const orgNameElement = xml.querySelector<XMLElement>('orgName');
-
-    return {
-      ...this.parseGenericEntity(xml),
-      label: (orgNameElement ? replaceNewLines(orgNameElement.textContent) : '') || 'No info',
-    };
-  }
-
-  private parseEvent(xml: XMLElement): NamedEntity {
-    const eventLabelElement = xml.querySelector<XMLElement>('label');
-
-    return {
-      ...this.parseGenericEntity(xml),
-      label: (eventLabelElement ? replaceNewLines(eventLabelElement.textContent) : '') || 'No info',
-    };
-  }
-
-  private parseRelation(xml: XMLElement): Relation {
-    const active = xml.getAttribute('active') || '';
-    const mutual = xml.getAttribute('mutual') || '';
-    const passive = xml.getAttribute('passive') || '';
-    const descriptionEls = xml.querySelectorAll<XMLElement>('desc');
-    const relation: Relation = {
-      type: Relation,
-      name: xml.getAttribute('name'),
-      activeParts: active.replace(/#/g, '').split(' '),
-      mutualParts: mutual.replace(/#/g, '').split(' '),
-      passiveParts: passive.replace(/#/g, '').split(' '),
-      relationType: xml.getAttribute('type'),
-      attributes: this.genericParserService.parseAttributes(xml),
-      content: Array.from(xml.children).map((subchild: XMLElement) => this.parseEntityInfo(subchild)),
-      description: [],
-    };
-    if (descriptionEls && descriptionEls.length > 0) {
-      descriptionEls.forEach((el) => (relation.description as Description).push(this.genericParserService.parse(el)));
-    } else {
-      relation.description = [this.genericParserService.parseText(xml)];
-    }
-    const parentListEl = xml.parentElement.tagName === 'listRelation' ? xml.parentElement : undefined;
-    if (parentListEl) {
-      relation.relationType = `${(parentListEl.getAttribute('type') || '')} ${(relation.relationType || '')}`.trim();
-    }
-
-    return relation;
-  }
-
-  private parseEntityInfo(xml: XMLElement): NamedEntityInfo {
-    return {
-      type: NamedEntityInfo,
-      label: xml.nodeType === 1 ? xml.tagName.toLowerCase() : 'info',
-      content: [this.genericParserService.parse(xml)],
-      attributes: xml.nodeType === 1 ? this.parseAttributes(xml) : {},
-    };
-  }
-
-  private parseAttributes(xml: XMLElement) {
-    const attributes: AttributesMap = {};
-    Array.from(xml.attributes).forEach((attr) => {
-      attributes[attr.name] = attr.value;
-    });
-
-    return attributes;
   }
 
   public getResultsByType(lists: NamedEntitiesList[], entities: NamedEntity[], type: string[]) {
@@ -265,11 +107,7 @@ export class NamedEntitiesParserService {
     return tagName.replace('list', '').toLowerCase();
   }
 
-  private getEntityType(tagName): NamedEntityType {
-    return tagName.toLowerCase();
-  }
-
-  public parseNamedEntitiesOccurrences(pages: PageData[]) {
+  public parseNamedEntitiesOccurrences(pages: Page[]) {
     return pages.map(p => this.getNamedEntitiesOccurrencesInPage(p))
       .reduce(
         (x, y) => {
@@ -286,7 +124,7 @@ export class NamedEntitiesParserService {
         {});
   }
 
-  public getNamedEntitiesOccurrencesInPage(p: PageData): Array<Map<NamedEntityOccurrence>> {
+  public getNamedEntitiesOccurrencesInPage(p: Page): Array<Map<NamedEntityOccurrence>> {
     return p.originalContent
       .filter(e => e.nodeType === 1)
       .map(e => {
@@ -327,10 +165,11 @@ export class NamedEntitiesParserService {
 
   private parseNamedEntityOccurrence(xml: XMLElement) {
     const doc = xml.closest('text');
+    const elementParser = createParser(ElementParser, parse);
 
     return {
       ref: xml.getAttribute('ref').replace('#', ''),
-      el: this.genericParserService.parseElement(xml),
+      el: elementParser.parse(xml),
       docId: doc ? doc.getAttribute('xml:id') : '', // TODO: get proper document id when missing
       docLabel: doc ? doc.getAttribute('n') || doc.getAttribute('xml:id') : '', // TODO: get proper document label when attributes missing
     };