+++ src/app/services/xml-parsers/witnesses-parser.service.ts
@@ -1,33 +1,29 @@
 import { Injectable } from '@angular/core';
-import { map, shareReplay } from 'rxjs/operators';
-import { Description, Witness, WitnessesData, WitnessGroup, XMLElement } from '../../models/evt-models';
+import { parse } from '.';
+import { Description, Witness, Witnesses, WitnessGroup, XMLElement } from '../../models/evt-models';
 import { isNestedInElem, xpath } from '../../utils/dom-utils';
 import { arrayToMap } from '../../utils/js-utils';
 import { replaceNotWordChar } from '../../utils/xml-utils';
-import { EditionDataService } from '../edition-data.service';
+import { AttributeParser } from './basic-parsers';
 import { GenericParserService } from './generic-parser.service';
+import { createParser, getID } from './parser-models';
 
 @Injectable({
   providedIn: 'root',
 })
 export class WitnessesParserService {
-  public readonly witnessesData$ = this.editionDataService.parsedEditionSource$.pipe(
-    map((source) => this.parseWitnessesData(source)),
-    shareReplay(1),
-  );
-
   private witListTagName = 'listWit';
   private witTagName = 'witness';
   private witNameAttr = 'type="siglum"';
   private groupTagName = 'head';
+  private attributeParser = createParser(AttributeParser, parse);
 
   constructor(
-    private editionDataService: EditionDataService,
     private genericParserService: GenericParserService,
   ) {
   }
 
-  private parseWitnessesData(document: XMLElement): WitnessesData {
+  public parseWitnessesData(document: XMLElement): Witnesses {
     const lists = Array.from(document.querySelectorAll<XMLElement>(this.witListTagName));
 
     return {
@@ -50,10 +46,12 @@ export class WitnessesParserService {
   }
 
   private parseWitness(wit: XMLElement): Witness {
+    const id = getID(wit);
+
     return {
-      id: wit.getAttribute('xml:id') || xpath(wit),
-      name: this.parseWitnessName(wit) || wit.getAttribute('xml:id') || xpath(wit),
-      attributes: this.genericParserService.parseAttributes(wit),
+      id,
+      name: this.parseWitnessName(wit) || id,
+      attributes: this.attributeParser.parse(wit),
       content: this.parseWitnessContent(wit),
       groupId: this.parseParentGroupId(wit),
     };
@@ -88,7 +86,7 @@ export class WitnessesParserService {
     return {
       id: list.getAttribute('xml:id') || xpath(list),
       name: this.parseGroupName(list) || replaceNotWordChar(list.getAttribute('xml:id')) || xpath(list),
-      attributes: this.genericParserService.parseAttributes(list),
+      attributes: this.attributeParser.parse(list),
       witnesses: this.parseGroupWitnesses(list),
       groupId: this.parseParentGroupId(list),
     };